# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  jeri_blog_internal_dto.AIAISummaryRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  jeri_blog_internal_dto.AIAISummaryResponse:
    properties:
      summary:
        description: 生成的AI摘要
        type: string
    type: object
  jeri_blog_internal_dto.AISummaryRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  jeri_blog_internal_dto.AISummaryResponse:
    properties:
      summary:
        description: 生成的摘要
        type: string
    type: object
  jeri_blog_internal_dto.AITestRequest:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      model:
        type: string
    required:
    - api_key
    - base_url
    - model
    type: object
  jeri_blog_internal_dto.AITitleRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  jeri_blog_internal_dto.AITitleResponse:
    properties:
      title:
        description: 生成的标题
        type: string
    type: object
...
  jeri_blog_internal_dto.FeedbackResponse:
    properties:
      admin_reply:
        type: string
      email:
        description: 联系邮箱
        type: string
      feedback_time:
        allOf:
        - $ref: '#/definitions/jeri_blog_pkg_utils.JSONTime'
        description: 反馈时间
      form_content:
        description: 表单内容
      id:
        type: integer
      ip:
        type: string
      reply_time:
        allOf:
        - $ref: '#/definitions/jeri_blog_pkg_utils.JSONTime'
        description: 回复时间
      report_type:
        type: string
      report_url:
        type: string
      status:
        type: string
      ticket_no:
        description: 工单号
        type: string
      user_agent:
        type: string
    type: object
...
  jeri_blog_internal_dto.SubmitFeedbackRequest:
    properties:
      attachmentFiles:
        description: 附件文件URL数组
        items:
          type: string
        type: array
      description:
        type: string
      email:
        description: 可选联系邮箱
        type: string
      reason:
        description: 可选原因字段
        type: string
      reportType:
        enum:
        - copyright
        - inappropriate
        - summary
        - suggestion
        type: string
      reportUrl:
        type: string
    required:
    - description
    - reportType
    - reportUrl
    type: object
...
  jeri_blog_internal_dto.UpdateFeedbackRequest:
    properties:
      admin_reply:
        type: string
      status:
        enum:
        - pending
        - resolved
        - closed
        type: string
    required:
    - status
    type: object
...
  jeri_blog_pkg_response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  jeri_blog_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  jeri_blog_pkg_utils.JSONTime:
    properties:
      time.Time:
        type: string
    type: object
info:
  contact:
    email: 416685476@qq.com
    name: Jerion
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Jeri-Server
  version: v1
paths:
  /admin/ai/ai-summary:
    post:
      consumes:
      - application/json
      description: 基于文章内容生成AI摘要（150-200字，旁观者角度）
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.AIAISummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.AIAISummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 生成AI摘要
      tags:
      - AI功能
  /admin/ai/summary:
    post:
      consumes:
      - application/json
      description: 基于文章内容自动生成摘要（50-100字，创作者角度）
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.AISummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.AISummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 生成文章摘要
      tags:
      - AI功能
  /admin/ai/test:
    post:
      consumes:
      - application/json
      description: 使用提供的配置测试AI服务连通性
      parameters:
      - description: AI配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.AITestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 测试AI配置
      tags:
      - AI功能
  /admin/ai/title:
    post:
      consumes:
      - application/json
      description: 根据内容生成多个标题建议
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.AITitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.AITitleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 生成标题建议
      tags:
      - AI功能
...
  /api/v1/admin/feedback:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/jeri_blog_pkg_response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/jeri_blog_internal_dto.FeedbackResponse'
                        type: array
                    type: object
              type: object
      summary: 获取反馈列表
      tags:
      - 反馈管理
  /api/v1/admin/feedback/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 反馈ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      summary: 删除反馈
      tags:
      - 反馈管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 反馈ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.FeedbackResponse'
              type: object
      summary: 获取反馈详情
      tags:
      - 反馈管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 反馈ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.UpdateFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      summary: 更新反馈
      tags:
      - 反馈管理
...
  /api/v1/feedback:
    post:
      consumes:
      - application/json
      parameters:
      - description: 反馈信息
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.SubmitFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.FeedbackResponse'
              type: object
      summary: 提交反馈投诉
      tags:
      - 反馈
  /api/v1/feedback/ticket/{ticket_no}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工单号
        in: path
        name: ticket_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.FeedbackResponse'
              type: object
      summary: 根据工单号查询反馈
      tags:
      - 反馈
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"